attendance filter moved new component
